refactor: merge storyFields.ts into profileFields.ts (β.10a)
storyFields was a 56-line file holding one constant array
(WOLF_BUILTIN_STORIES) and a derived Set. Per-reviewer feedback (and
correctly noted by the user), the structural shape between stories
and experience/project/education is identical — both are
arrays-of-tables keyed by stable id. The reason storyFields existed
as its own file was the wolf-managed builtin registry, not the shape.

Since `profileFields.ts` already holds wolf-managed schema metadata
(PROFILE_FIELDS / REQUIRED_PROFILE_FIELDS), the builtin story
registry belongs in the same file. One source of truth for "what
wolf hard-codes about the profile schema".

# Changes

- Move WOLF_BUILTIN_STORIES + WOLF_BUILTIN_STORY_IDS + BuiltinStory
  type into profileFields.ts.
- Delete src/utils/storyFields.ts.
- Update 6 import sites:
  - src/runtime/migrations/v1ToV2.ts
  - src/runtime/migrations/__tests__/v1ToV2.test.ts
  - src/utils/profileToml.ts
  - src/utils/__tests__/profileToml.test.ts
  - src/application/impl/profileApplicationServiceImpl.ts
  - src/application/impl/doctorApplicationServiceImpl.ts
- Coalesce the two profileFields imports into one statement per file
  (cosmetic, post-rename cleanup).

# Behaviour unchanged

Builtin protections (`wolf profile remove story <id>` rejects
builtins; `wolf profile set story.<id>.prompt|required` rejects
builtins) still work — they import WOLF_BUILTIN_STORY_IDS from the
new location, same Set, same lookup. Verified by the existing tests
in profileApplicationServiceImpl.test.ts (and full suite).

Total suite: 346 / 346 passing.
